# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Path AI API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export interface SignupInput {
  name: string;
  email: string;
  password: string;
}

export interface LoginInput {
  email: string;
  password: string;
}

export interface User {
  id: string;
  name: string;
  email: string;
  onboardingComplete?: boolean;
  createdAt: string;
}

export interface AuthResponse {
  user: User;
  token: string;
}

export interface UserProfileInput {
  educationLevel: string;
  course: string;
  school?: string;
  state: string;
  academicLevel?: string;
  interests: string[];
  existingSkills?: string[];
  monthlyBudget: string;
  deviceAccess?: string;
  internetQuality?: string;
  careerGoals?: string;
  learningPace: string;
  workStyle?: string;
  timeline: string;
}

export interface UserProfile {
  id: string;
  userId: string;
  educationLevel: string;
  course: string;
  /** @nullable */
  school?: string | null;
  state: string;
  /** @nullable */
  academicLevel?: string | null;
  interests: string[];
  existingSkills?: string[];
  monthlyBudget: string;
  /** @nullable */
  deviceAccess?: string | null;
  /** @nullable */
  internetQuality?: string | null;
  /** @nullable */
  careerGoals?: string | null;
  learningPace: string;
  /** @nullable */
  workStyle?: string | null;
  timeline: string;
}

export interface CareerPath {
  id: string;
  title: string;
  description: string;
  salaryMin: number;
  salaryMax: number;
  demandLevel: string;
  remotePotential: string;
  difficultyLevel: string;
  timeToJobReady: string;
  requiredSkills: string[];
  category?: string;
}

export interface CareerPathDetail {
  id: string;
  title: string;
  description: string;
  salaryMin: number;
  salaryMax: number;
  demandLevel: string;
  remotePotential: string;
  difficultyLevel: string;
  timeToJobReady: string;
  requiredSkills: string[];
  tools: string[];
  responsibilities: string[];
  entryStrategy: string;
  category?: string;
}

export interface CareerRecommendation {
  career: CareerPath;
  fitScore: number;
  whyItFits: string;
  recommendedNextAction: string;
}

export interface SkillGapAnalysis {
  careerId: string;
  careerTitle: string;
  readinessScore: number;
  skillsHave: string[];
  skillsMissing: string[];
  priorityOrder: string[];
  estimatedTimeToClose: string;
}

export interface RoadmapInput {
  careerId: string;
}

export interface RoadmapTask {
  id: string;
  roadmapId: string;
  title: string;
  description: string;
  stage: string;
  /** @nullable */
  monthNumber?: number | null;
  /** @nullable */
  resourceLink?: string | null;
  estimatedDuration: string;
  difficulty: string;
  status: string;
  /** @nullable */
  completedAt?: string | null;
}

export interface Roadmap {
  id: string;
  userId: string;
  careerId: string;
  careerTitle: string;
  generatedAt: string;
  tasks: RoadmapTask[];
}

export interface TaskCompletionInput {
  completed: boolean;
}

export interface ChatMessage {
  id: string;
  role: string;
  content: string;
  createdAt: string;
}

export interface ChatMessageInput {
  content: string;
}

export interface ProgressSummary {
  totalTasks: number;
  completedTasks: number;
  completionPercentage: number;
  currentStage: string;
  /** @nullable */
  currentCareerPath: string | null;
  tasksCompletedThisWeek: number;
  streakDays: number;
  upcomingMilestones?: string[];
}

export interface Resource {
  id: string;
  title: string;
  type: string;
  url: string;
  isFree: boolean;
  provider: string;
  estimatedDuration: string;
  relatedSkill: string;
  category: string;
  difficulty?: string;
}

export type ListResourcesParams = {
skill?: string;
category?: string;
};