# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Career Path AI API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/signup:
    post:
      operationId: signup
      tags: [auth]
      summary: Create a new account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignupInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Validation error
  /auth/login:
    post:
      operationId: login
      tags: [auth]
      summary: Log in
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
...
  /auth/me:
    get:
      operationId: getMe
      tags: [auth]
      summary: Get current user
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
        "401":
          description: Unauthorized
...
  /careers/{id}:
    get:
      operationId: getCareer
      tags: [careers]
      summary: Get career path detail
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CareerPathDetail"
        "404":
          description: Not found
...
  /careers/{careerId}/skill-gap:
    get:
      operationId: getSkillGap
      tags: [careers]
      summary: Get skill-gap analysis for a career path
      parameters:
        - name: careerId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SkillGapAnalysis"
...
components:
  schemas:
...
    SignupInput:
      type: object
      required: [name, email, password]
      properties:
        name:
          type: string
        email:
          type: string
        password:
          type: string
    LoginInput:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
        password:
          type: string
    User:
      type: object
      required: [id, name, email, createdAt]
      properties:
        id:
          type: string
        name:
          type: string
        email:
          type: string
        onboardingComplete:
          type: boolean
        createdAt:
          type: string
    AuthResponse:
      type: object
      required: [user, token]
      properties:
        user:
          $ref: "#/components/schemas/User"
        token:
          type: string
...
    CareerPathDetail:
      type: object
      required: [id, title, description, salaryMin, salaryMax, demandLevel, remotePotential, difficultyLevel, timeToJobReady, requiredSkills, tools, responsibilities, entryStrategy]
      properties:
        id:
          type: string
        title:
          type: string
        description:
          type: string
        salaryMin:
          type: integer
        salaryMax:
          type: integer
        demandLevel:
          type: string
        remotePotential:
          type: string
        difficultyLevel:
          type: string
        timeToJobReady:
          type: string
        requiredSkills:
          type: array
          items:
            type: string
        tools:
          type: array
          items:
            type: string
        responsibilities:
          type: array
          items:
            type: string
        entryStrategy:
          type: string
        category:
          type: string
...
    SkillGapAnalysis:
      type: object
      required: [careerId, careerTitle, readinessScore, skillsHave, skillsMissing, priorityOrder, estimatedTimeToClose]
      properties:
        careerId:
          type: string
        careerTitle:
          type: string
        readinessScore:
          type: integer
        skillsHave:
          type: array
          items:
            type: string
        skillsMissing:
          type: array
          items:
            type: string
        priorityOrder:
          type: array
          items:
            type: string
        estimatedTimeToClose:
          type: string
...