# lib/api-zod/src/generated/types/careerPathDetail.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Path AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CareerPathDetail {
  id: string;
  title: string;
  description: string;
  salaryMin: number;
  salaryMax: number;
  demandLevel: string;
  remotePotential: string;
  difficultyLevel: string;
  timeToJobReady: string;
  requiredSkills: string[];
  tools: string[];
  responsibilities: string[];
  entryStrategy: string;
  category?: string;
}