# lib/api-zod/src/generated/types/skillGapAnalysis.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * Career Path AI API specification
 * OpenAPI spec version: 0.1.0
 */

export interface SkillGapAnalysis {
  careerId: string;
  careerTitle: string;
  readinessScore: number;
  skillsHave: string[];
  skillsMissing: string[];
  priorityOrder: string[];
  estimatedTimeToClose: string;
}